+++ week2-mission1/src/main.jsx
@@ -1,8 +1,10 @@
-
 import { createRoot } from 'react-dom/client'
 import App from './App.jsx'
 import './index.css'
+import { TodoContextProvider } from './context/TodoContext.jsx'
 
 createRoot(document.getElementById('root')).render(
-    <App />
+    <TodoContextProvider>
+        <App/>
+    </TodoContextProvider>
 )
